+++ app/upgrades/v9/accounts.go
@@ -1,4 +1,4 @@
-package v83
+package v9
 
 var Accounts = [1810][2]string{
 	{"evmos1009egsf8sk3puq3aynt8eymmcqnneezkkvceav", "20299404928815863552"},
